% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_permanent_snow.R
\name{find_date_permanent_snow}
\alias{find_date_permanent_snow}
\title{Find start date of permanent snow in a winter}
\usage{
find_date_permanent_snow(
  date,
  snow_depth,
  plot = TRUE,
  temp.window.size = 14,
  endrule = c("min", "NA", "trim", "keep", "constant", "func")[1],
  align = c("left", "center", "right")[1],
  set_start = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{date}{vector. Date of the weather records.}

\item{snow_depth}{vector. Snow depth values.}

\item{plot}{logical. Whether the plot of the start date of permanent snow should be plotted
or not. Default is `TRUE`.}

\item{temp.window.size}{numeric. Size of the temporal window to calculate the minimum
amount of snow, in days. Default is 14 days.}

\item{endrule}{character. String indicating how the values at the beginning and
the end, of the array, should be treated. Default is "min". For more options,
see the function \code{\link[caTools]{runmin}}).}

\item{align}{character. How should the temporal window be aligned?
Default is "left". For more information, see \code{\link[caTools]{runmin}}).}

\item{set_start}{POSIX object. Date and time of the start of the winter.
Used mainly for debugging, to test other dates. Default is NULL.}

\item{debug}{logical. If debugging, plot also the snow depth values after
minimum filtering. Default is FALSE.}

\item{...}{}
}
\value{
Date object. Start date of the permanent snow in a winter.
}
\description{
This function uses a time series of snow depth to find the beginning
of the permanent snow in a given winter. This is based on calculating the
minimum amount of snow in a window of a given temporal length to check if,
after a given snow event, there was melting and the snow vanished.
}
\examples{
library(dplyr)
data(weather_mittadalen)

weather_2019 <- weather_mittadalen \%>\%
  dplyr::filter(year == 2019)

with(weather_2019, find_date_permanent_snow(date, snow_depth, temp.window.size = 14))

# not run
with(weather_2019, find_date_permanent_snow(date, snow_depth, temp.window.size = 14,
                                            set_start = lubridate::ymd_h("2018-11-01 00")))

}
