% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_winter_events.R
\name{identify_winter_event3}
\alias{identify_winter_event3}
\alias{identify_winter_event4}
\title{Identify harsh weather winter events}
\usage{
identify_winter_event3(
  date,
  temp_min,
  temp_max,
  precipitation,
  temp_min_thr = -2,
  temp_max_thr = 2,
  prec_thr = 3
)

identify_winter_event4(
  date,
  temp_avg,
  precipitation,
  temp_avg_day = +1,
  temp_avg_next_day = -1,
  prec_thr = 3
)
}
\arguments{
\item{date}{date object. vector. Vector of dates in a time series of a given winter.}

\item{temp_min}{numeric. vector. Minimum daily temperature, in Celsius.}

\item{temp_max}{numeric. vector. Minimum daily temperature, in Celsius.}

\item{precipitation}{numeric. vector. Precipitation accumulated each day, in mm.}

\item{temp_min_thr}{numeric. Threshold of minimum temperature, in Celsius, used for event 3.}

\item{temp_max_thr}{numeric. Threshold of maximum temperature, in Celsius, used for event 3.}

\item{prec_thr}{numeric. Threshold of precipitation, in mm, used for events 3 and 4.}

\item{temp_avg}{numeric. vector. Average daily temperature, in Celsius.}

\item{temp_avg_day}{numeric. Threshold of average temperature in the day, in Celsius, used for event 4.}

\item{temp_avg_next_day}{numeric. Threshold of average temperature in the following day, in Celsius, used for event 4.}
}
\value{
A list with the elements:
\describe{
  \item{events}{time series with 0 for days with no event and 1 with days when the event occurred.}
  \item{n_days}{number of days when the event occurred.}
  \item{n_events}{number of events, by aggregating subsequent days when it repeatedly happened.}
  \item{duration_events}{duration of each event, in days.}
  \item{event_dates}{starting date of each event.}
  \item{n_event_date}{time series with the counter of the number of events. Similar to events.}
} These elements are calculated for both the whole winter and only the first half of the winter.
}
\description{
This functions identifies the winter events related to temperature and precipitation described in
Lundqvist et al. (2007). The functions uses daily temperature (min, max, or average, in degrees Celcius) and
precipitation (in mm) to identify events that may be harsh weather conditions for reindeer husbandry.
}
\details{
The events are:
  * Events 1 and 2 are still to be implemented;
  * Event 3: Day when max temp > 2, min temp < -2, and precipitation > 3 mm;
  * Event 4: Day when average temp > 1, precipitation > 3 mm, followed by day with average temp < -1.
}
\examples{
library(ggplot2)
library(dplyr)
data(weather_mittadalen)

weather <- weather_mittadalen \%>\%
  dplyr::filter(year == 2014)

# events 3
with(weather, identify_winter_event3(date, temp_min, temp_max, prec))

# events 4
with(weather, identify_winter_event4(date, temp_avg, prec))

# visualize events 4 for this year
weather_condition <-
  with(weather, analyze_weather(date, snow_depth, prec, temp_min, temp_max,
                                     temp_avg, start = "first_date",
                                     plot_first_snow = T))

to_plot <- weather_condition

to_plot$weather_indices <- weather_condition$weather_indices \%>\%
  dplyr::mutate(temp_avg_next = c(temp_avg[-1], NA)) \%>\%
  dplyr::filter(lubridate::month(date) == 12)

to_plot \%>\%
  plot_weather(term = c("temp_avg", "prec", "temp_avg_next"), factor_mult = c(1,1)) +
  geom_hline(yintercept = c(-1,1), linetype = 2) +
  ylim(-10, 5)
}
\references{
Lundqvist, H., Norell, L., Danel, O. 2007. Multivariate characterisation of environmental
conditions for reindeer husbandry in Sweden. Rangifer, 27(1): 5-23.
}
