% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation_mittadalen.R
\docType{data}
\name{weather_mittadalen}
\alias{weather_mittadalen}
\title{Winter weather data for Mittadalen}
\format{A data frame with 7 variables:
\describe{
  \item{date}{date of the record}
  \item{year}{year of the record}
  \item{snow_depth}{snow depth in mm, measured in Myskelasen}
  \item{prec}{snow depth in mm, measured in Dravagen}
  \item{temp_min}{minimum daily temperature in degrees Celcius, measured in Dravagen}
  \item{temp_max}{maximum daily temperature in degrees Celcius, measured in Dravagen}
  \item{temp_avg}{average daily temperature in degrees Celcius, measured in Dravagen}
}}
\source{
SMHI
}
\usage{
data(weather_mittadalen)
data(weather_mittadalen_allstations)
}
\description{
Winter weather data for Mittadalen Sami reindeer herding district in Sweden.
The dataset presents daily data for weather conditions from the winter 2008 to 2020.
Winter here comprises from beginning of October of the previous year to last of Aril
of the current year. A second dataset presents data for more weather stations
in the region.
}
\examples{
library(ggplot2)
data(weather_mittadalen)

ggplot(weather_mittadalen, aes(date)) +
  facet_wrap(.~year, scales = "free", ncol = 5)+
  geom_line(aes(y = snow_depth, color = "Snow depth (mm)"))  +
  geom_line(aes(y = prec, color = "Precipitation (mm)")) +
  labs(x = "Date", y = "", color = "Measure") +
  theme(legend.position = "bottom")

data(weather_tassasen)

ggplot(weather_tassasen, aes(date)) +
  facet_wrap(.~year, scales = "free", ncol = 5)+
  geom_line(aes(y = snow_depth, color = "Snow depth (mm)"))  +
  geom_line(aes(y = prec, color = "Precipitation (mm)")) +
  labs(x = "Date", y = "", color = "Measure") +
  theme(legend.position = "bottom")
}
\keyword{conditions}
\keyword{data,}
\keyword{weather}
